requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.24.1

python-dotenv==1.0.0
python-docx==0.8.11
reportlab==4.0.9
supabase==2.0.0         # supabase-py - ensure compatibility with your environment
pydantic==1.10.11
